Project/requirements.txt
flask
Flask-SQLAlchemy
flask-login
PyJWT==1.6.4
